# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fake Mart API Documentation",
        "description": "This is the API documentation for my Laravel app.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/branch/create": {
            "post": {
                "tags": [
                    "Branch"
                ],
                "summary": "Create a new branch",
                "description": "Creates a new branch with name, location, and contact number.",
                "operationId": "createBranch",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "contact_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Main Branch"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Phnom Penh"
                                    },
                                    "contact_number": {
                                        "type": "string",
                                        "example": "012345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "new_data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Main Branch"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Phnom Penh"
                                                },
                                                "contact_number": {
                                                    "type": "string",
                                                    "example": "012345678"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-06T12:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-06T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "The name field is required."
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "The location field is required."
                                                },
                                                "contact_number": {
                                                    "type": "string",
                                                    "example": "The contact number field is required."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/branch/update": {
            "post": {
                "tags": [
                    "Branch"
                ],
                "summary": "Update a branch by ID",
                "description": "Updates a branch with the provided ID, name, location, and contact number.",
                "operationId": "updateBranch",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "location",
                                    "contact_number"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Branch Name"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Siem Reap"
                                    },
                                    "contact_number": {
                                        "type": "string",
                                        "example": "098765432"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "update_data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Updated Branch Name"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Siem Reap"
                                                },
                                                "contact_number": {
                                                    "type": "string",
                                                    "example": "098765432"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-06T12:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-06T12:10:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "resource not found"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/branch/delete": {
            "post": {
                "tags": [
                    "Branch"
                ],
                "summary": "Delete a branch by ID (POST version)",
                "description": "Deletes a branch if the given ID exists. This uses POST instead of DELETE for compatibility reasons.",
                "operationId": "postDeleteBranch",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch deleted successfully or resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "delete_data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Main Branch"
                                                        },
                                                        "location": {
                                                            "type": "string",
                                                            "example": "Phnom Penh"
                                                        },
                                                        "contact_number": {
                                                            "type": "string",
                                                            "example": "012345678"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2025-08-06T12:00:00.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "example": "2025-08-06T12:00:00.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "status_code": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "resource not found"
                                                },
                                                "status_code": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}